+++ src/Bullet/App.php
@@ -15,6 +15,19 @@ class App extends Container
         $this->currentCallbacks = &$this->callbacks;
     }
 
+    protected function executeCallback(\Closure $c, $request)
+    {
+        $c = \Closure::bind($c, $this);
+        $response = $c($request);
+
+        if ($response === null || $response instanceOf Response) {
+            return $response;
+        }
+        if (is_string($response)) {
+            return new Response($response, 200);
+        }
+    }
+
     /**
      * Run app with given Request
      *
@@ -28,10 +41,8 @@ class App extends Container
         $this->currentCallbacks = &$this->callbacks;
 
         try {
-            $rsp = new Response();
-
             // Remove empty path elements
-            $uri = $request->uri();
+            $uri = $request->url();
             $parts = [];
             foreach (explode('/', $uri) as $part) {
                 if ($part != '') {
@@ -43,22 +54,45 @@ class App extends Container
 
             // TODO: run before filter
 
-            $level = 0;
+            // Walk through the URI and execute path callbacks
             foreach ($parts as $part) {
-                if (!array_key_exists($part, $this->currentCallbacks)) {
-                    print "404\n";
-                    print implode('/', array_slice($parts, 0, $level+1));
-                    throw new NotFoundException();
+                // Try to find a callback array for the current URI part
+                if (!array_key_exists($part, $this->currentCallbacks) || !array_key_exists('path', $this->currentCallbacks[$part])) {
+                    return new Response(null, 404);
+                }
+
+                // Yes, there is a callback array, set it as the current one
+                $this->currentCallbacks = $this->currentCallbacks[$part];
+
+                // Execute path callback
+                $rsp = $this->executeCallback($this->currentCallbacks['path'], $request);
+
+                // If there's already a response, return it and finish parsing the URL
+                if ($rsp instanceOf Response) {
                     return $rsp;
                 }
-                if (array_key_exists('path', $this->currentCallbacks[$part])) {
-                    $cb = Closure::bind($this->currentCallbacks[$part]['path'], $this);
-                    $cb();
-                }
-                ++$level;
             }
 
-            return $rsp;
+            $method = $request->method();
+
+            // The URI has been processed. Call the appropriate method callback
+            if (!array_key_exists($method, $this->currentCallbacks)) {
+                // Nope, we can't serve this URI, 405 Not Allowed
+                return new Response(null, 405);
+            }
+
+            // There indeed is a method callback, so let's call it!
+            $rsp = $this->executeCallback($this->currentCallbacks[$method], $request);
+
+            // If there's a response, we can return it
+            if ($rsp instanceOf Response) {
+                return $rsp;
+            }
+
+            // TODO: formats?
+            //return new Response(406); // Not acceptable format
+
+            return new Response(null, 501); // Got no error, but got no response either. This is "Not Implemented".
         } finally {
             $this->currentCallbacks = &$currentCallbacks;
         }
@@ -74,39 +108,39 @@ class App extends Container
         $this->currentCallbacks[$part]['path'] = $callback;
     }
 
-    public function param($part, $callback)
+    public function param($param, $callback)
     {
-        $this->currentCallbacks[$part]['path'] = $callback;
+        $this->currentCallbacks[$part]['param'] = $callback;
     }
 
-    public function get($part, $callback)
+    public function get($callback)
     {
-        $this->currentCallbacks[$part]['get'] = $callback;
+        $this->currentCallbacks['GET'] = $callback;
     }
 
-    public function post($part, $callback)
+    public function post($callback)
     {
-        $this->currentCallbacks[$part]['post'] = $callback;
+        $this->currentCallbacks['POST'] = $callback;
     }
 
-    public function put($part, $callback)
+    public function put($callback)
     {
-        $this->currentCallbacks[$part]['put'] = $callback;
+        $this->currentCallbacks['PUT'] = $callback;
     }
 
-    public function delete($part, $callback)
+    public function delete($callback)
     {
-        $this->currentCallbacks[$part]['delete'] = $callback;
+        $this->currentCallbacks['DELETE'] = $callback;
     }
 
-    public function patch($part, $callback)
+    public function patch($callback)
     {
-        $this->currentCallbacks[$part]['patch'] = $callback;
+        $this->currentCallbacks['PATCH'] = $callback;
     }
 
-    public function options($part, $callback)
+    public function options($callback)
     {
-        $this->currentCallbacks[$part]['options'] = $callback;
+        $this->currentCallbacks['OPTIONS'] = $callback;
     }
 
     public function domain()